* master: (123 commits)
  Update affected tests to use timezone aware datetime objects.
  Temporary disable the check.
  Fix lint.
  Improve tzinfo check.
  Typos.
  Fix lint.
  Update changelog.
  Throw if the value passed in contains an invalid timezone.
  Add UTC timezone info to the resulting value, update affected tests.
  Fix broken test and add a new one.
  Fix a bug with ComplexDateTimeField - make sure we are always operating with UTC.
  Fixes on Kami's comments.
  Test creation of right kind of sensors
  None config is acceptable for PollingSensors
  Use issubclass instead of isinstance check
  Add another constant.
  Update docs.
  Include additional common environment variables in the actions environment.
  Remove now obsolete environment variable - this has been unified and moved into new ST2_ACTION_AUTH_TOKEN environment variable.
  Fix docs build issues.
  ...

+++ contrib/packs/actions/pack_mgmt/download.py
@@ -1,6 +1,23 @@
+# Licensed to the StackStorm, Inc ('StackStorm') under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 import shutil
 import hashlib
+import re
+import json
 
 import six
 from git.repo import Repo
@@ -11,21 +28,41 @@ from st2actions.runners.pythonrunner import Action
 ALL_PACKS = '*'
 PACK_REPO_ROOT = 'packs'
 MANIFEST_FILE = 'pack.yaml'
+CONFIG_FILE = 'config.yaml'
+GITINFO_FILE = '.gitinfo'
 PACK_RESERVE_CHARACTER = '.'
 
+STACKSTORM_CONTRIB_REPOS = [
+    'st2contrib',
+    'st2incubator'
+]
 
-class InstallGitRepoAction(Action):
-    def run(self, packs, repo_url, abs_repo_base, verifyssl=True, branch='master'):
-        repo_name = repo_url[repo_url.rfind('/') + 1: repo_url.rfind('.')]
+
+class DownloadGitRepoAction(Action):
+    def __init__(self, config=None):
+        super(DownloadGitRepoAction, self).__init__(config=config)
+        self._subtree = None
+        self._repo_url = None
+
+    def run(self, packs, repo_url, abs_repo_base, verifyssl=True, branch='master', subtree=False):
+        self._subtree = self._eval_subtree(repo_url, subtree)
+        self._repo_url = self._eval_repo_url(repo_url)
+
+        repo_name = self._repo_url[self._repo_url.rfind('/') + 1: self._repo_url.rfind('.')]
         lock_name = hashlib.md5(repo_name).hexdigest() + '.lock'
 
         with LockFile('/tmp/%s' % (lock_name)):
-            abs_local_path = self._clone_repo(repo_url=repo_url, verifyssl=verifyssl,
+            abs_local_path = self._clone_repo(repo_url=self._repo_url, verifyssl=verifyssl,
                                               branch=branch)
             try:
-                # st2-contrib repo has a top-level packs folder that actually contains the
-                pack_abs_local_path = os.path.join(abs_local_path, PACK_REPO_ROOT)
-                result = self._move_packs(abs_repo_base, packs, pack_abs_local_path)
+                if self._subtree:
+                    # st2-contrib repo has a top-level packs folder that actually contains the
+                    pack_abs_local_path = os.path.join(abs_local_path, PACK_REPO_ROOT)
+                else:
+                    pack_abs_local_path = abs_local_path
+
+                self._tag_pack(pack_abs_local_path, packs, self._subtree)
+                result = self._move_packs(abs_repo_base, packs, pack_abs_local_path, self._subtree)
             finally:
                 self._cleanup_repo(abs_local_path)
         return self._validate_result(result=result, packs=packs, repo_url=repo_url)
@@ -40,24 +77,29 @@ class InstallGitRepoAction(Action):
         # Disable SSL cert checking if explictly asked
         if not verifyssl:
             os.environ['GIT_SSL_NO_VERIFY'] = 'true'
-
-        Repo.clone_from(repo_url, abs_local_path, branch=branch)
+        # Shallow clone the repo to avoid getting all the metadata. We only need HEAD of a
+        # specific branch so save some download time.
+        Repo.clone_from(repo_url, abs_local_path, branch=branch, depth=1)
         return abs_local_path
 
-    def _move_packs(self, abs_repo_base, packs, abs_local_path):
+    def _move_packs(self, abs_repo_base, packs, abs_local_path, subtree):
         result = {}
         # all_packs should be removed as a pack with that name is not expected to be found.
         if ALL_PACKS in packs:
             packs = os.listdir(abs_local_path)
         for pack in packs:
-            abs_pack_temp_location = os.path.join(abs_local_path, pack)
-            desired, message = InstallGitRepoAction._is_desired_pack(abs_pack_temp_location, pack)
+            abs_pack_temp_location = os.path.join(abs_local_path, pack) if subtree else abs_local_path
+            desired, message = DownloadGitRepoAction._is_desired_pack(abs_pack_temp_location, pack)
             if desired:
                 to = abs_repo_base
                 dest_pack_path = os.path.join(abs_repo_base, pack)
                 if os.path.exists(dest_pack_path):
                     self.logger.debug('Removing existing pack %s in %s to replace.', pack,
                                       dest_pack_path)
+                    # Ensure to preserve any existing configuration
+                    old_config_file = os.path.join(dest_pack_path, CONFIG_FILE)
+                    new_config_file = os.path.join(abs_pack_temp_location, CONFIG_FILE)
+                    shutil.move(old_config_file, new_config_file)
                     shutil.rmtree(dest_pack_path)
                 self.logger.debug('Moving pack from %s to %s.', abs_pack_temp_location, to)
                 shutil.move(abs_pack_temp_location, to)
@@ -88,7 +130,8 @@ class InstallGitRepoAction(Action):
     @staticmethod
     def _cleanup_repo(abs_local_path):
         # basic lock checking etc?
-        shutil.rmtree(abs_local_path)
+        if os.path.isdir(abs_local_path):
+            shutil.rmtree(abs_local_path)
 
     @staticmethod
     def _validate_result(result, packs, repo_url):
@@ -115,3 +158,44 @@ class InstallGitRepoAction(Action):
             raise Exception(message)
 
         return sanitized_result
+
+    @staticmethod
+    def _eval_subtree(repo_url, subtree):
+        match = False
+        for stackstorm_repo_name in STACKSTORM_CONTRIB_REPOS:
+            if stackstorm_repo_name in repo_url:
+                match = True
+                break
+
+        return subtree | match
+
+    @staticmethod
+    def _eval_repo_url(repo_url):
+        """Allow passing short GitHub style URLs"""
+        has_git_extension = repo_url.endswith('.git')
+        if len(repo_url.split('/')) == 2 and "git@" not in repo_url:
+            url = "https://github.com/{}".format(repo_url)
+        else:
+            url = repo_url
+        return url if has_git_extension else "{}.git".format(url)
+
+    @staticmethod
+    def _tag_pack(pack_dir, packs, subtree):
+        """Add git information to pack directory for retrieval later"""
+        for pack in packs:
+            repo = Repo(pack_dir)
+            payload = {
+                "branch": repo.active_branch.name,
+                "ref": repo.head.commit.hexsha
+            }
+
+            if subtree:
+                info_file = os.path.join(pack_dir, pack, GITINFO_FILE)
+            else:
+                info_file = os.path.join(pack_dir, GITINFO_FILE)
+
+            try:
+                gitinfo = open(info_file, "w")
+                gitinfo.write(json.dumps(payload))
+            finally:
+                gitinfo.close()
